[Silver III] Title: 주유소, Time: 508 ms, Memory: 56276 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.kt
@@ -1,9 +1,9 @@
 fun main() {
     val city = readln().toInt()
-    val distanceBetweenCity = readln().split(" ").map { it.toInt() }
-    val oilPrice = readln().split(" ").map { it.toInt() }
+    val distanceBetweenCity = readln().split(" ").map { it.toLong() }
+    val oilPrice = readln().split(" ").map { it.toLong() }
 
-    var totalPay = 0
+    var totalPay = 0L
     var currentMin = oilPrice.first()
 
     oilPrice.zip(distanceBetweenCity).forEach { (price, distance) ->
